KSAP-49 convert variable credits to BigDecimal
git-svn-id: https://svn.kuali.org/repos/student/enrollment/ks-api/branches/KSAP@54945 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-ap-api/src/main/java/org/kuali/student/myplan/academicplan/infc/PlanItem.java
@@ -1,11 +1,11 @@
 package org.kuali.student.myplan.academicplan.infc;
 
+import java.math.BigDecimal;
+import java.util.List;
+
 import org.kuali.student.r2.common.infc.HasId;
 import org.kuali.student.r2.common.infc.RichText;
 import org.kuali.student.r2.common.infc.TypeStateEntity;
-
-import java.util.List;
-import java.util.Set;
 
 /**
  * A single plan item. Plan items can link to different entities in the system based on the refObjectType.
@@ -48,6 +48,6 @@ public interface PlanItem extends HasId, TypeStateEntity {
     * The plan credits
     * @name The number of credits
     */
-    public Float getCredit();
+    public BigDecimal getCredit();
 
 }
